+++ src/extra/background.cpp
@@ -9,7 +9,7 @@ SAD_DECLARE(Background,Sprite)
 Background::Background(const hst::string &name,const hRectF & r): 
 Sprite(
 	   sad::TextureManager::instance()->get(name),
-	   hst::rect< ::s3d::point>(::s3d::point(0,SCREEN_HEIGHT,0),::s3d::point(SCREEN_WIDTH,0,0)),
+	   hst::rect< ::s3d::point>(::s3d::point(0.0f,(float)SCREEN_HEIGHT,0.0f),::s3d::point((float)SCREEN_WIDTH,0.0f,0.0f)),
 	   r)
 {
 }
@@ -17,7 +17,7 @@ Sprite(
 Background::Background(const hst::string &name): 
 Sprite(
 	   TEX_PTR,
-	   hst::rect< ::s3d::point>(::s3d::point(0,SCREEN_HEIGHT,0),::s3d::point(SCREEN_WIDTH,0,0)),
+	   hst::rect< ::s3d::point>(::s3d::point(0.0f,(float)SCREEN_HEIGHT,0.0f),::s3d::point((float)SCREEN_WIDTH,0.0f,0.0f)),
 	   hRectF(hPointF(0,0),hPointF(TEX_PTR->width(),TEX_PTR->height())))
 {
 }
